Move positionMixin to positionStoreMixin
