+++ src/components/yBar.js
@@ -11,7 +11,7 @@ import PropTypes from 'prop-types';
 import createRef from 'create-react-ref/lib/createRef';
 import createShallowCompare from '../utils/createShallowCompare';
 
-import { positionStoreMixin } from '../store/positionReducers';
+import positionStoreMixin from '../store/positionStoreMixin';
 
 class ListComponent extends PureComponent {
